backend/requirements.txt
flask==2.0.1
requests==2.26.0
python-whois==0.7.3
beautifulsoup4==4.10.0
python-dotenv==0.19.0
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
flask-bcrypt==0.7.1
authlib==1.5.1
requests-oauthlib==2.0.0
gunicorn==20.1.0